create route and controller for image upload
